refactor: 클래스를 테스트 파일에서 main 으로 추출했다.
